requirements.txt
Flask
Gunicorn
pandas>=1.2.0
python-docx==0.8.11
requests==2.28.2
